data - Use public data_t API in data_convert_tree()
_convert_tree() now declares its match parameter as a data_type_t, so
the int cast that hid the internal type_t it used to take is no longer
needed.

No functional change.

+++ src/common/data.c
@@ -2058,7 +2058,7 @@ static size_t _convert_tree(data_t *data, const data_type_t match)
 
 extern size_t data_convert_tree(data_t *data, const data_type_t match)
 {
-	return _convert_tree(data, (int) match);
+	return _convert_tree(data, match);
 }
 
 static data_for_each_cmd_t _find_dict_match(const char *key, const data_t *a,
